requirements.txt
Django==4.2.5
djangorestframework==3.14.0
drf-yasg==1.21.7
django-model-utils==4.3.1